Memoized SectionInfo component
